extract common code into a function
