+++ code/lib/builder-manager/src/index.ts
@@ -21,8 +21,10 @@ import {
 } from './types';
 import { getData } from './utils/data';
 import { safeResolve } from './utils/safeResolve';
+import { readOrdererFiles } from './utils/files';
 
-let compilation: Compilation;
+// eslint-disable-next-line import/no-mutable-exports
+export let compilation: Compilation;
 let asyncIterator: ReturnType<StarterFunction> | ReturnType<BuilderFunction>;
 
 export const getConfig: ManagerBuilder['getConfig'] = async (options) => {
@@ -118,17 +120,9 @@ const starter: StarterFunction = async function* starterGeneratorFn({
   router.use(`/sb-addons`, express.static(addonsDir));
   router.use(`/sb-manager`, express.static(coreDirOrigin));
 
-  const files = await Promise.all(
-    compilation?.outputFiles?.map(async (file) => {
-      await ensureFile(file.path).then(() => writeFile(file.path, file.contents));
-      return file.path.replace(addonsDir, './sb-addons');
-    }) || []
-  );
+  const { cssFiles, jsFiles } = await readOrdererFiles(addonsDir);
 
   yield;
-
-  const jsFiles = files.filter((file) => file.endsWith('.mjs'));
-  const cssFiles = files.filter((file) => file.endsWith('.css'));
 
   const html = await renderHTML(
     template,
@@ -191,17 +185,9 @@ const builder: BuilderFunction = async function* builderGeneratorFn({ startTime,
   yield;
 
   const managerFiles = copy(coreDirOrigin, coreDirTarget);
-  const files = await Promise.all(
-    compilation?.outputFiles?.map(async (file) => {
-      await ensureFile(file.path).then(() => writeFile(file.path, file.contents));
-      return file.path.replace(addonsDir, './sb-addons');
-    }) || []
-  );
+  const { cssFiles, jsFiles } = await readOrdererFiles(addonsDir);
 
   yield;
-
-  const jsFiles = files.filter((file) => file.endsWith('.mjs'));
-  const cssFiles = files.filter((file) => file.endsWith('.css'));
 
   const html = await renderHTML(
     template,
